requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
sentence-transformers>=2.2.0
ultralytics>=8.0.0
faiss-cpu>=1.7.4
Pillow>=10.0.0
transformers>=4.30.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
inference-sdk>=0.50.0
inference[transformers]
inference[sam]
inference[clip]
inference[gaze]
inference[grounding-dino]
inference[yolo-world]
supervision>=0.3.0 